started work on bot1, final changes to 1-player dealer to come
The bot uses the same logic as the dealer (hit on soft 17), will
implement the same to the dealer next along with tracking wins, loses
and reporting them to the user. I should probably add that I currently
have no idea whether this code works, so do with that what you will.